Separated some structs in styling and css3 crates
