+++ Cargo.toml
@@ -44,6 +44,7 @@ gosub_net = { path = "./crates/gosub_net", features = [] }
 gosub_config = { path = "./crates/gosub_config", features = [] }
 gosub_html5 = { path = "./crates/gosub_html5", features = [] }
 gosub_css3 = { path = "./crates/gosub_css3", features = [] }
+gosub_styling = { path = "./crates/gosub_styling", features = [] }
 gosub_webexecutor = { path = "./crates/gosub_webexecutor", features = [] }
 gosub_jsapi = { path = "./crates/gosub_jsapi", features = [] }
 gosub_testing = { path = "./crates/gosub_testing", features = [] }
