+++ crates/gosub_css3/Cargo.toml
@@ -10,4 +10,5 @@ gosub_shared = { path = "../gosub_shared", features = [] }
 lazy_static = "1.4"
 log = "0.4.20"
 simple_logger = "4.2.0"
+anyhow = { version = "1.0.80", features = [] }
