+++ crates/gosub_css3/src/lib.rs
@@ -4,10 +4,12 @@ use crate::parser_config::{Context, ParserConfig};
 use crate::tokenizer::Tokenizer;
 use gosub_shared::byte_stream::{ByteStream, Encoding, Stream};
 
+pub mod convert;
 pub mod location;
-pub mod node;
-mod parser;
+mod node;
+pub mod parser;
 pub mod parser_config;
+pub mod stylesheet;
 pub mod tokenizer;
 mod unicode;
 pub mod walker;
