+++ crates/gosub_css3/src/parser_config.rs
@@ -8,7 +8,7 @@ pub enum Context {
     Declaration,
 }
 
-/// ParserConfig holds the configuration for the parser
+/// ParserConfig holds the configuration for the CSS3 parser
 pub struct ParserConfig {
     /// Context defines how the data needs to be parsed
     pub context: Context,
@@ -16,7 +16,8 @@ pub struct ParserConfig {
     pub location: Location,
     /// Optional source filename or url
     pub source: Option<String>,
-    /// Ignore errors and continue parsing
+    /// Ignore errors and continue parsing. Any errors will not be returned in the final AST
+    /// (this means if a selector is invalid, all rules will be ignored, even when they are valid)
     pub ignore_errors: bool,
 }
