+++ crates/gosub_html5/Cargo.toml
@@ -7,8 +7,12 @@ license = "MIT"
 
 [dependencies]
 gosub_shared = { path = "../gosub_shared", features = [] }
+gosub_css3 = { path = "../gosub_css3", features = [] }
 derive_more = "0.99"
 phf = { version = "0.11.2", features = ["macros"] }
 lazy_static = "1.4"
 thiserror = "1.0.57"
+url = { version = "2.5.0", features = [] }
+log = { version = "0.4.20", features = [] }
+ureq = "2.9.6"
