+++ crates/gosub_html5/src/parser/document.rs
@@ -10,9 +10,10 @@ use crate::parser::query::{Condition, Query};
 use crate::parser::quirks::QuirksMode;
 use crate::parser::tree_builder::TreeBuilder;
 use crate::util::is_valid_id_attribute_value;
-use crate::styles::styles::CssStylesheet;
+use crate::visit::Visitor;
 use core::fmt;
 use core::fmt::Debug;
+use gosub_css3::stylesheet::CssStylesheet;
 use gosub_shared::types::Result;
 use std::cell::RefCell;
 use std::collections::HashMap;
