+++ crates/gosub_styling/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "gosub_styling"
+version = "0.1.0"
+edition = "2021"
+authors = ["Gosub Community <info@gosub.io>"]
+license = "MIT"
+
+[dependencies]
+gosub_shared = { path = "../gosub_shared" }
+gosub_css3 = { path = "../gosub_css3" }
+gosub_html5 = { path = "../gosub_html5" }
+lazy_static = "1.4"
+anyhow = "1.0.80"
