+++ crates/gosub_styling/src/lib.rs
@@ -1,0 +1,8 @@
+//! Styling functionality
+//!
+//! This crate connects CSS3 and HTML5 into a styling pipeline
+//!
+
+pub mod calculator;
+pub mod css_colors;
+pub mod pipeline;
