+++ crates/gosub_styling/src/shorthands.rs
