+++ src/bin/style-parser.rs
@@ -1,18 +1,18 @@
 use anyhow::{bail, Result};
-use gosub_engine::bytes::{CharIterator, Confidence, Encoding};
-use gosub_engine::html5::node::data::comment::CommentData;
-use gosub_engine::html5::node::data::doctype::DocTypeData;
-use gosub_engine::html5::node::data::document::DocumentData;
-use gosub_engine::html5::node::data::element::ElementData;
-use gosub_engine::html5::node::data::text::TextData;
-use gosub_engine::html5::node::Node;
-use gosub_engine::html5::parser::document::DocumentBuilder;
-use gosub_engine::html5::parser::document::{visit, Document};
-use gosub_engine::html5::parser::Html5Parser;
-use gosub_engine::html5::visit::Visitor;
-use gosub_engine::pipeline::Pipeline;
-use gosub_engine::styles::css_colors::get_color_value;
-use gosub_engine::styles::StyleCalculator;
+use gosub_html5::node::data::comment::CommentData;
+use gosub_html5::node::data::doctype::DocTypeData;
+use gosub_html5::node::data::document::DocumentData;
+use gosub_html5::node::data::element::ElementData;
+use gosub_html5::node::data::text::TextData;
+use gosub_html5::node::Node;
+use gosub_html5::parser::document::DocumentBuilder;
+use gosub_html5::parser::document::{visit, Document};
+use gosub_html5::parser::Html5Parser;
+use gosub_html5::visit::Visitor;
+use gosub_shared::bytes::{CharIterator, Confidence, Encoding};
+use gosub_styling::calculator::StyleCalculator;
+use gosub_styling::css_colors::get_color_value;
+use gosub_styling::pipeline::Pipeline;
 use regex::Regex;
 use std::collections::HashMap;
 use std::fs;
